+++ playground/backend/login.py
@@ -10,7 +10,7 @@ import streamlit as st
 from cryptography.hazmat.primitives.kdf.pbkdf2 import PBKDF2HMAC
 from cryptography.hazmat.primitives import hashes
 from cryptography.fernet import Fernet
-from ui.configuration import configuration
+from playground.configuration import configuration
 
 logger = logging.getLogger(__name__)
