+++ pyproject.toml
@@ -19,7 +19,7 @@ dependencies = [
 ]
 
 [project.optional-dependencies]
-ui = [
+playground = [
     "streamlit==1.45.1",
     "streamlit-extras==0.5.0", 
     "bcrypt==4.3.0",
